Move class related likelihood from test.py to framework.py.
