--- logs/README.md
@@ -1,1 +1,0 @@
-# logs here
